requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.2.0
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
coverage==7.4.4
cryptography==42.0.5
distlib==0.3.9
Django==5.0.6
django-filter==24.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-docstrings==1.7.0
flake8-quotes==3.4.0
gitdb==4.0.11
GitPython==3.1.43
identify==2.6.1
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.6.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==71.0.4
smmap==5.0.1
snowballstemmer==2.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.27.1
